+++ docs/source/contrib.autoguide.rst
@@ -1,0 +1,9 @@
+Automatic Guide Generation
+--------------------------
+
+.. automodule:: pyro.contrib.autoguide.advi
+    :members:
+    :undoc-members:
+    :show-inheritance:
+    :member-order: bysource
+
