+++ docs/source/inference_algos.rst
@@ -29,15 +29,6 @@ ELBO
     :undoc-members:
     :show-inheritance:
 
-ADVI
-----
-
-.. automodule:: pyro.infer.advi
-    :members:
-    :undoc-members:
-    :show-inheritance:
-    :member-order: bysource
-
 Importance
 ----------
