+++ pyro/contrib/autoguide/__init__.py
@@ -1,0 +1,5 @@
+from __future__ import absolute_import, division, print_function
+
+from pyro.contrib.autoguide.advi import ADVIDiagonalNormal, ADVIDiscreteParallel, ADVIMaster, ADVIMultivariateNormal
+
+# flake8: noqa
