+++ pyro/infer/__init__.py
@@ -1,7 +1,6 @@
 from __future__ import absolute_import, division, print_function
 
 from pyro.infer.abstract_infer import Marginal, TracePosterior
-from pyro.infer.advi import ADVIDiagonalNormal, ADVIDiscreteParallel, ADVIMaster, ADVIMultivariateNormal
 from pyro.infer.elbo import ELBO
 from pyro.infer.enum import config_enumerate
 from pyro.infer.importance import Importance
